autotest/requirements.txt
pytest>=7.2.0
requests>=2.28.0
allure-pytest>=2.12.0
pydantic>=1.10.0
PyYAML>=6.0
jsonschema>=4.17.0
faker>=15.0.0
pytest-html>=3.1.0
pytest-xdist>=3.1.0
pytest-rerunfailures>=11.0